+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/clang/__init__.py
@@ -1,1 +1,1 @@
-from build_system.cmd.compiler.host.get_info.version.clang.fetch_clang_version import fetch
+from build_system.cmd.compiler.host.get_info.version.clang.impl import fetch
