+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/clang/impl.py
